[refactor] #78 - 구독 커스텀훅 생성
